Tidies up provider and local path usage
The intent here is to get a module that behaves in a way that terraform 0.12 users would more readily expect. In addition, region configuration is handled by the AWS provider configuration we inherit, as opposed to by some other external definition. Currently availability zone definitions are still coming from "the environment", which should perhaps be changed to an input in the future.